OR-531 fix compilation issues

+++ src/main/java/ca/bestbuy/orders/fraud/dao/FraudRequestStatusHistoryRepository.java
@@ -50,5 +50,5 @@ public interface FraudRequestStatusHistoryRepository extends OrderFraudBaseRepos
 	 * @param requestVersion
 	 * @return
 	 */
-	 Iterable<FraudRequestStatusHistory> findByFraudRequestOrderNumberAndFraudRequestRequestVersion(BigDecimal orderNumber, BigDecimal requestVersion, Sort sort);
+	 Iterable<FraudRequestStatusHistory> findByFraudRequestOrderNumberAndFraudRequestRequestVersion(BigDecimal orderNumber, Long requestVersion, Sort sort);
 }
